# openapi.yaml
...
openapi: 3.1.0
info:
  title: SMIRK API
  version: "1.0.0"
  description: Route inventory generated from concrete Express route declarations in server.ts and src/team-routes.ts. Middleware app.use declarations are excluded from paths.
servers:
  - url: https://smirkcalls.com
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    WorkspaceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Workspace session or invite token
    ProvisioningBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Provisioning token
paths:
...
  "/api/agents":
    get:
      operationId: "GET_api_agents"
      tags:
        - "agents"
      summary: "GET /api/agents"
      x-source: "src/routes/agent-routes.ts:23"
      x-express-path: "/api/agents"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    post:
      operationId: "POST_api_agents"
      tags:
        - "agents"
      summary: "POST /api/agents"
      x-source: "src/routes/agent-routes.ts:30"
      x-express-path: "/api/agents"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/agents/{id}":
    delete:
      operationId: "DELETE_api_agents_id"
      tags:
        - "agents"
      summary: "DELETE /api/agents/:id"
      x-source: "src/routes/agent-routes.ts:76"
      x-express-path: "/api/agents/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    get:
      operationId: "GET_api_agents_id"
      tags:
        - "agents"
      summary: "GET /api/agents/:id"
      x-source: "src/routes/agent-routes.ts:92"
      x-express-path: "/api/agents/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    patch:
      operationId: "PATCH_api_agents_id"
      tags:
        - "agents"
      summary: "PATCH /api/agents/:id"
      x-source: "src/routes/agent-routes.ts:101"
      x-express-path: "/api/agents/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    put:
      operationId: "PUT_api_agents_id"
      tags:
        - "agents"
      summary: "PUT /api/agents/:id"
      x-source: "src/routes/agent-routes.ts:53"
      x-express-path: "/api/agents/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/agents/{id}/activate":
    put:
      operationId: "PUT_api_agents_id_activate"
      tags:
        - "agents"
      summary: "PUT /api/agents/:id/activate"
      x-source: "src/routes/agent-routes.ts:44"
      x-express-path: "/api/agents/:id/activate"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/agents/active":
    get:
      operationId: "GET_api_agents_active"
      tags:
        - "agents"
      summary: "GET /api/agents/active"
      x-source: "src/routes/agent-routes.ts:84"
      x-express-path: "/api/agents/active"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...